requirements.txt
# Core LangChain dependencies
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langchain-openai==0.3.7
langchain-together==0.3.7
langchain-text-splitters==0.3.0
langchain-pinecone==0.2.0

# LLM & Embedding providers
openai>=1.3.5
together>=0.2.9

# Vector stores
pinecone-client>=3.0.0
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
tiktoken>=0.5.1
tqdm==4.66.1
PyYAML==6.0

# ML & Data processing
transformers>=4.36.2
sentence-transformers>=2.2.2
scikit-learn==1.3.2
numpy>=1.26.2
pandas==2.2.2

# Dataset management
datasets==2.13.1

# CLI & UI
rich>=13.7.0
streamlit==1.28.0
